test: fix netpoll ut in windows
