+++ pkg/app/client/client_test.go
@@ -48,7 +48,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"math/rand"
 	"net"
 	"net/http"
 	"net/http/httptest"
@@ -70,7 +69,6 @@ import (
 	"github.com/cloudwego/hertz/pkg/common/test/assert"
 	"github.com/cloudwego/hertz/pkg/network"
 	"github.com/cloudwego/hertz/pkg/network/dialer"
-	"github.com/cloudwego/hertz/pkg/network/netpoll"
 	"github.com/cloudwego/hertz/pkg/network/standard"
 	"github.com/cloudwego/hertz/pkg/protocol"
 	"github.com/cloudwego/hertz/pkg/protocol/consts"
@@ -1897,20 +1895,6 @@ func (m *mockDialer) DialConnection(network, address string, timeout time.Durati
 	return m.Dialer.DialConnection(m.network, m.address, m.timeout, tlsConfig)
 }
 
-func newMockDialerWithCustomFunc(network, address string, timeout time.Duration, f func(network, address string, timeout time.Duration, tlsConfig *tls.Config)) network.Dialer {
-	dialer := standard.NewDialer()
-	if rand.Intn(2) == 0 {
-		dialer = netpoll.NewDialer()
-	}
-	return &mockDialer{
-		Dialer:           dialer,
-		customDialerFunc: f,
-		network:          network,
-		address:          address,
-		timeout:          timeout,
-	}
-}
-
 func TestClientRetry(t *testing.T) {
 	t.Parallel()
 	client, err := NewClient(
@@ -2045,12 +2029,12 @@ func TestClientDialerName(t *testing.T) {
 		t.Errorf("expected 'netpoll', but get %s", dName)
 	}
 
-	client, _ = NewClient(WithDialer(netpoll.NewDialer()))
+	client, _ = NewClient(WithDialer(&mockDialer{}))
 	dName, err = client.GetDialerName()
 	if err != nil {
 		t.Fatalf("unexpected error: %v", err)
 	}
-	if dName != "netpoll" {
+	if dName != "client" {
 		t.Errorf("expected 'standard', but get %s", dName)
 	}
