+++ pkg/network/netpoll/dial.go
@@ -13,9 +13,6 @@
 // limitations under the License.
 //
 
-//go:build !windows
-// +build !windows
-
 package netpoll
 
 import (
